rename(#4): entity 위치 변경
