+++ service/payment/server/src/main/java/com/sparta/payment/domain/entity/PaymentState.java
@@ -1,4 +1,4 @@
-package com.sparta.payment.domain;
+package com.sparta.payment.domain.entity;
 
 public enum PaymentState {
   // 결제 완료, 결제 대기, 결제 실패, 결제 취소, 환불 대기, 환불 완료
